Change Structure
This aims to simplify the setup and structure of the config. It should
now be more organised and less full in the parent directory as all
configs now live in .homedir. By doing this the setup was also reduced
to one command. The Readme was adapted. Aliases to edit configuration
files were adapted accordingly. This will probably mean that the
symlinks have to be renewed on all devices as the config files changed
location. Otherwise no changes were made.